Factor out floating column.
