+++ source/index.ts
@@ -1,4 +1,5 @@
 export * from './array_table_model';
+export * from './floating_column';
 export * from './operations';
 export * from './table_model';
 export * from './table_view';
